Hibakezelés try,catch

+++ haromszogek/Haromszog.cs
@@ -6,11 +6,11 @@ namespace haromszogek
 {
     internal class Haromszog
     {
-        private int aoldal;
-        private int boldal;
-        private int coldal;
+        private double aoldal;
+        private double boldal;
+        private double coldal;
 
-        public Haromszog(int aoldal, int boldal, int coldal)
+        public Haromszog(double aoldal, double boldal, double coldal)
         {
             this.aoldal = aoldal;
             this.boldal = boldal;
